Module Kcas.Mode

Operating modes of the k-CAS-n-CMP algorithm.

type t = [

| `Lock_free
(*
In `Lock_free mode the algorithm makes sure that at least one domain will be able to make progress at the cost of performing read-only operations as read-write operations.
*)

| `Obstruction_free
(*
In `Obstruction_free mode the algorithm proceeds optimistically and allows read-only operations to fail due to interference from other domains that might have been prevented in the `Lock_free mode.
*)

]

Type of an operating mode of the k-CAS-n-CMP algorithm.